third_party/blink/web_tests/dom/document_type/previous-element-sibling.html

<!DOCTYPE html>
<title>This tests the NonDocumentTypeChildNode interface previousElementSibling property on a DocumentType node</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var doctype = document.doctype;
assert_equals(doctype.previousElementSibling, undefined);
done();
</script>
